Split models into files
